# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gerenciamento de Consultas",
        "description": "Documentação da API para o projeto de agendamentos e serviços.",
        "contact": {
            "email": "seu.email@dominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desenvolvimento Local"
        }
    ],
    "paths": {
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Lista todas as consultas (paginado)",
                "operationId": "listAppointments",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de consultas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AppointmentsResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Cria uma nova consulta",
                "operationId": "createAppointment",
                "requestBody": {
                    "description": "Dados da consulta a ser criado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAppointmentsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consulta criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AppointmentsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Busca uma consulta pelo ID",
                "operationId": "getAppointmentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agendamento a ser buscado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agendamento encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AppointmentsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consulta não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Atualiza uma consulta existente",
                "operationId": "updateAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da consulta a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados da consulta a ser atualizado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAppointmentsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "consulta atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AppointmentsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "consulta não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/appointments/cancel/{id}": {
            "patch": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Cancela uma consulta",
                "operationId": "cancelAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da consulta a ser cancelada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Consulta cancelado com sucesso"
                    },
                    "404": {
                        "description": "consulta não encontrada"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registra um novo usuário",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterAuthRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação (ex: e-mail já existe)"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Autentica um usuário e retorna um token",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginAuthRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthResponse": {
                "title": "Auth Response",
                "description": "Resposta de sucesso de Login ou Registro",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "token": {
                        "type": "string",
                        "example": "1|aBcDeFgHiJkLmNoPqRsTuVwXyZ123456"
                    }
                },
                "type": "object"
            },
            "LoginAuthRequest": {
                "title": "Login Auth Request",
                "description": "Payload para login de usuário",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "usuario@exemplo.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "senha123"
                    }
                },
                "type": "object"
            },
            "RegisterAuthRequest": {
                "title": "Register Auth Request",
                "description": "Payload para registro de novo usuário",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Weslley"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "weslley@exemplo.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "senha123"
                    }
                },
                "type": "object"
            },
            "StoreAppointmentsRequest": {
                "title": "Store Appointments Request",
                "description": "Payload para criar uma nova consulta",
                "required": [
                    "title",
                    "scheduled_at",
                    "user_id",
                    "services"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Consulta de Rotina"
                    },
                    "scheduled_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-30T14:30:00Z"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
...
            "UpdateAppointmentsRequest": {
                "title": "Update Appointments Request",
                "description": "Payload para atualizar um agendamento",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Consulta de Retorno"
                    },
                    "scheduled_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-15T10:00:00Z"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "confirmed"
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
            "AppointmentsResource": {
                "title": "Appointments Resource",
                "description": "Representação de um agendamento",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Consulta de Rotina"
                    },
                    "scheduled_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "10/10/2025 10:40:30"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "total_price": {
                        "type": "string",
                        "format": "decimal",
                        "example": "150.50"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ServicesResource"
                        }
                    }
                },
                "type": "object"
            },
            "ServicesResource": {
                "description": "Recurso que representa um Serviço",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Corte de cabelo"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "description": "Representação de um usuário",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Weslley"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "weslley@exemplo.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum (obrigatório para endpoints protegidos)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}